+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## [0.0.1] - Icon resize animation
+## [0.0.2] - Add finising widget and child option
+
+* When loop has finished, it can show a widget. `SizedBox.shrink()` may be used to hide the icon.
+* It can display any widgets instead of icon or combination of icon and widgets.
+* ## [0.0.1] - Icon resize animation
 
 * First verion